Add id to key export and call key import/export from restore/backup
Restoring the user id ensures that USAGs still work as expected,
provided that the user id is restored before the groups.

This commit adds --with-user-sql to call the key export/import
functions from the backup and restore commands.  Sequel is required
very late so the tool can still be used without it if desired.