Remove special handling of billing-admin group; only update admins once
As far as I can tell, there is not particular reason to upload the
billing-admin's group or acl separately from the rest of the groups.

Also, the final update of the admin's group is unnecessary as that
should be correct as it is included in the objects upload by copy_to.