[ec_backup] Remove @error variable
We were not consistently using this, so its value was rather dubious.
I would love for this to be a better command line tool that returned a
sensible return code, but currently the common failure mode in this
tool is to raise an error.  Thus, all this was doing was cluttering
the code a bit.