Upload groups twice to work around group POST behavior in Chef 12
Chef 12 Server ignores group members as part of the POST content for a
new group.  On the second upload, ChefFS will use a PUT rather than a
POST to update the existing group with the proper membership.  This
also means that the topological sorting is uncessary on Chef 12, but
it is still required for Chef 11 suport.