Restore clients to correct org by querying org_id from stored org_name
The org_id contained in the backup is likely incorrect on the new
server. To restore client keys correctly, we need export the org_name
as part of the key data and then use the org_name on restore to find
the org_id.

A simple in-memory cache is used to avoid multiple org_id lookpups for
organizations with many clients.