Swallow errors when restoring files that 404
Because knife-ec-backup does not capture Chef Server API version
metadata when backing up, its impossible to know which files to
restore. This commit fixes a bug where a backup/restore occuring prior
to the introduction of a certain file (in this case
public_key_read_access) would fail because that file (rightfully) does
not exist.

There should be a follow up to this commit to add better logic so that
we can know which files we need to backup/restore, rather than simply
recovering from attempts to restore files that don't exist.

Signed-off-by: Tom Duffield <tom@chef.io>