Update Gemfile.lock
We failed to update Gemfile.lock with the most recent version bump,
which means it was being re-resolved during the omnibus-build in Chef
Server.

We are solving this in Chef Server by adding --local to the install
for now.

I think in the long run we want to remove this lockfile completely
since knife-ec-backup and its deps should be locked at a higher level
in the application that depends on it.

Signed-off-by: Steven Danna <steve@chef.io>