Remove Gemfile.lock
In addition to being released as a gem, knife-ec-backup is included in
the Chef Server build. Unfortunately, gem version management in Chef
Server is haphazard. Currently, knife-ec-backup is installed with
`--local`, which requires that all of its dependencies already be
installed on the machine. The goal of this is to prevent it from
inadvertently upgrading dependencies that may be used in the
chef-server-ctl commands.

The previous update to Gemfile.lock bumped many of our dependencies to
Chef 13, which is not yet shipped in Chef Server, breaking the build.

Removing the Gemfile.lock will likely allow the Chef Server build to
complete as bundler will use the existing versions installed in Chef
Server.

If you are thinking, "this is horrible," I agree with you.  The actual
solution is to more sensibly manage the dependencies of
chef-server-ctl so that we don't have to jump through hoops to avoid
unintentional pollution.

If you are thinking, "if you agree, then why are we doing this," the
answer is that we want to unblock the Chef Server pipeline without
re-writing how much of the chef-server-ctl build works at the moment.

Signed-off-by: Steven Danna <steve@chef.io>