[SUSTAIN-657] Handle and report user restore fails
Currently, importing user data on a restore can fail with a less than
optimal error message, and kill the restore altogether.

This change rescues these failures, adds some color to the exception
message about the probable cause, and adds them the error collection to
be reported at the end of the run.

Signed-off-by: Nolan Davidson <ndavidson@chef.io>